Moved invite button from Header to Compare Panel, added displaying email alongside name for users shared with
