+++ src/components/HeaderActionBar/index.tsx
@@ -4,12 +4,9 @@ import {
   faCalendarAlt,
   faPaste,
   faCaretDown,
-  faShare,
-  faCircle,
 } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import React, { useCallback, useContext, useState } from 'react';
-import useLocalStorageState from 'use-local-storage-state';
+import React from 'react';
 
 import { Button } from '..';
 import {
@@ -17,11 +14,10 @@ import {
   LARGE_DESKTOP_BREAKPOINT,
 } from '../../constants';
 import useMedia from '../../hooks/useMedia';
-import { AccountContext, AccountContextValue } from '../../contexts/account';
+import { AccountContextValue } from '../../contexts/account';
 import { classes } from '../../utils/misc';
 import { DropdownMenu, DropdownMenuAction } from '../Select';
 import AccountDropdown from '../AccountDropdown';
-import InvitationModal from '../InvitationModal';
 import ShareIcon from '../ShareIcon';
 
 import './stylesheet.scss';
@@ -37,9 +33,6 @@ export type HeaderActionBarProps = {
   onDownloadCalendar?: () => void;
   enableDownloadCalendar?: boolean;
 };
-
-// Key to mark when a user has already seen the invite modal.
-const MODAL_LOCAL_STORAGE_KEY = '2023-05-10-spr2023-invite-modal';
 
 /**
  * Displays the icon buttons (with optional text)
@@ -59,25 +52,6 @@ export default function HeaderActionBar({
   onDownloadCalendar = (): void => undefined,
   enableDownloadCalendar = false,
 }: HeaderActionBarProps): React.ReactElement {
-  const { type } = useContext(AccountContext);
-
-  const [invitationOpen, setInvitationOpen] = useState(false);
-  const [seenInviteModal, setSeenInviteModal] = useLocalStorageState<boolean>(
-    MODAL_LOCAL_STORAGE_KEY,
-    {
-      defaultValue: false,
-      storageSync: true,
-    }
-  );
-
-  const openInvitation = useCallback(() => {
-    setInvitationOpen(true);
-    if (!seenInviteModal) {
-      setSeenInviteModal(true);
-    }
-  }, [seenInviteModal, setSeenInviteModal]);
-  const hideInvitation = useCallback(() => setInvitationOpen(false), []);
-
   // Coalesce the export options into the props for a single <DropdownMenu>
   const enableExport =
     enableCopyCrns || enableDownloadCalendar || enableExportCalendar;
@@ -127,23 +101,6 @@ export default function HeaderActionBar({
           <FontAwesomeIcon fixedWidth icon={faCaretDown} />
         </div>
       </DropdownMenu>
-      <InvitationModal show={invitationOpen} onHide={hideInvitation} />
-
-      <Button
-        onClick={openInvitation}
-        disabled={type === 'signedOut'}
-        className={classes('header-action-bar__button', 'invite-button')}
-      >
-        <FontAwesomeIcon
-          className="header-action-bar__button-icon"
-          fixedWidth
-          icon={faShare}
-        />
-        <div className="header-action-bar__button-text">Invite</div>
-        {seenInviteModal || type === 'signedOut' ? null : (
-          <FontAwesomeIcon className="circle" fixedWidth icon={faCircle} />
-        )}
-      </Button>
 
       <Button
         href="https://github.com/gt-scheduler/website"
